+++ changelog.md
@@ -1,3 +1,8 @@
+# Version 2023.10.1 (2023-10-2)
+
+- Use enum for JsonRPC methods
+- Get supported json rpc methods and check against used methods
+
 # Version 2023.10.0 (2023-10-1)
 
 - Code cleanup